Completing task 1
checkout function now computes the total number of items and then
the price as a sum of items: n_item * price_item.
In doing so it checks the offers list for each item, if there is an
offer on the item it calls the offer function to compute the offered
price.